+++ sql/hive-thriftserver/src/test/scala/org/apache/spark/sql/hive/thriftserver/HiveSessionImplSuite.scala
@@ -17,8 +17,6 @@
 package org.apache.spark.sql.hive.thriftserver
 
 import java.lang.reflect.InvocationTargetException
-import java.nio.ByteBuffer
-import java.util.UUID
 
 import scala.collection.JavaConverters._
 import scala.collection.mutable
@@ -27,7 +25,6 @@ import org.apache.hadoop.hive.conf.HiveConf
 import org.apache.hive.service.cli.OperationHandle
 import org.apache.hive.service.cli.operation.{GetCatalogsOperation, Operation, OperationManager}
 import org.apache.hive.service.cli.session.{HiveSession, HiveSessionImpl, SessionManager}
-import org.apache.hive.service.rpc.thrift.{THandleIdentifier, TOperationHandle, TOperationType}
 
 import org.apache.spark.SparkFunSuite
 
@@ -65,31 +62,6 @@ class HiveSessionImplSuite extends SparkFunSuite {
   }
 }
 
-class GetCatalogsOperationMock(parentSession: HiveSession)
-  extends GetCatalogsOperation(parentSession) {
-
-  override def runInternal(): Unit = {}
-
-  override def getHandle: OperationHandle = {
-    val uuid: UUID = UUID.randomUUID()
-    val tHandleIdentifier: THandleIdentifier = new THandleIdentifier()
-    tHandleIdentifier.setGuid(getByteBufferFromUUID(uuid))
-    tHandleIdentifier.setSecret(getByteBufferFromUUID(uuid))
-    val tOperationHandle: TOperationHandle = new TOperationHandle()
-    tOperationHandle.setOperationId(tHandleIdentifier)
-    tOperationHandle.setOperationType(TOperationType.GET_TYPE_INFO)
-    tOperationHandle.setHasResultSetIsSet(false)
-    new OperationHandle(tOperationHandle)
-  }
-
-  private def getByteBufferFromUUID(uuid: UUID): Array[Byte] = {
-    val bb: ByteBuffer = ByteBuffer.wrap(new Array[Byte](16))
-    bb.putLong(uuid.getMostSignificantBits)
-    bb.putLong(uuid.getLeastSignificantBits)
-    bb.array
-  }
-}
-
 class OperationManagerMock extends OperationManager {
   private val calledHandles: mutable.Set[OperationHandle] = new mutable.HashSet[OperationHandle]()
 
@@ -114,3 +86,4 @@ class OperationManagerMock extends OperationManager {
 
   def getCalledHandles: mutable.Set[OperationHandle] = calledHandles
 }
+
